services/login/pyproject.toml
[project]
name = "login"
version = "0.1.0"
description = "Login service"
#authors = [{name = "", email = ""}]
requires-python = ">= 3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-dateutil",
    "python-dotenv",
    "pydantic",
    "pymongo",
    "redis",
    "pika",
    "loguru",
    "nicegui",
]

[project.urls]
Homepage = "https://example.com"

[build-system]
requires = ["setuptools>=61.2"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
debug = [
  "psycopg-binary",
  "remote-pdb",
]

[tool.uv]
dev-dependencies = [
  "pytest",
  "pytest-randomly",
  "black",
  "flake8",
  "mypy",
]

[tool.setuptools.packages.find]
where = ["source"]
namespaces = false

